requirements.txt
PyNaCl==1.5.0
discord.py==2.3.2
regex==2024.5.15
gTTS==2.5.0
instaloader==4.14.1
matplotlib==3.7.5
pyaml==23.9.7
pydub==0.25.1
requests==2.31.0
zipfile36==0.1.3
feedparser==6.0.11
googlesearch-python==1.2.3
beautifulsoup4==4.12.3
fuzzywuzzy[speedup]==0.18.0
chatterbot==1.2.3
chatterbot_corpus==1.2.2
SQLAlchemy==2.0.39
spacy==3.8.4
llama-index==0.12.34
llama-index-llms-gemini==0.4.14